% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catosmamenities-package.R
\docType{package}
\name{catosmamenities-package}
\alias{catosmamenities}
\alias{catosmamenities-package}
\title{catosmamenities: Download amenities from official Catalan sources and compare with OSM}
\description{
This package downloads certain types of amenities both from OpenStreetMap and from official sources from the Catalan government or local authorities in Catalonia in order to compare it, identify missing objects or complete existing ones.
}
\author{
\strong{Maintainer}: Marc Bosch Matas \email{mboschmatas@gmail.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
