% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{correspondences}
\alias{correspondences}
\title{Correspondence of amenity categories as defined by the Catalan government and OSM tags}
\format{
A tibble with 340 rows and 5 variables
}
\usage{
data(correspondences)
}
\description{
A dataset containing every type of amenity as detected by he Catalan government linking it to one or several key/Value OSM pairs
}
\details{
\itemize{
\item categoria_1. Main category of amenity
\item categoria_2. Secondary category of amenity.
\item categoria_3. Optional tertiary category of amenity
\item key. Corresponding OSM key
\item value. Corresponding OSM value.
}
}
\keyword{datasets}
