% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_schools.R
\name{get_all_schools}
\alias{get_all_schools}
\title{get_all_schools}
\usage{
get_all_schools(place, plot = TRUE)
}
\arguments{
\item{place}{character. name of the place for which to get schools - municipality or comarca}

\item{plot}{logical. If true,returns a list with an interactive tmap plot with both sources plotted and a data.frame. If false, only a data.frame}
}
\value{
A tibble with all schools or a list with said tibble and a map.
}
\description{
get all schools registered in openstreetmapand the Catalan government from a municipality or comarca in Cataloniamat
}
