% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gencat_schools.R
\name{get_gencat_schools}
\alias{get_gencat_schools}
\title{get_gencat_schools}
\usage{
get_gencat_schools(place, is_sf = TRUE)
}
\arguments{
\item{place}{character. name of the place for which to get schools - municipality or comarca}

\item{is_sf}{logical. If true, returns sf table. If false, a tibble.}
}
\value{
A sf or tibble with all schools in the desired place.
}
\description{
get all schools from the Catalan Government overpass from a municipality or comarca in Catalonia in either sf or table format
}
